[Chore] revert to production configs
